fix(lint): split ui/routines/state_tests.rs under the 500-line gate
